Put the address under the map, deleted old code, and actually populating address

+++ London Pub Crawl/LPCLineOptionModalViewController.m
@@ -86,7 +86,7 @@ LPCLine *selectedLine;
 }
 
 - (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section {
-    int number = [self.filteredStationArray count];
+    long number = [self.filteredStationArray count];
     return number;
 }
